Fix bedrock2 wrapper functions
This is kind-of hacky

Note that we now also insert casts on the return value.  This probably
isn't required, but it seems better to do it, and I was previously not
doing it only because I didn't have easy access to the type.

+++ fiat-bedrock2/src/p384_64.c
@@ -2767,7 +2767,7 @@ uintptr_t internal_fiat_p384_nonzero(uintptr_t in0) {
 
 /* NOTE: The following wrapper function is not covered by Coq proofs */
 static void fiat_p384_nonzero(uint64_t* out1, const uint64_t arg1[6]) {
-  *out1 = internal_fiat_p384_nonzero((uintptr_t)arg1);
+  *out1 = (uint64_t)internal_fiat_p384_nonzero((uintptr_t)arg1);
 }
