+++ fiat-bedrock2/src/secp256k1_32.c
@@ -4535,7 +4535,7 @@ uintptr_t internal_fiat_secp256k1_nonzero(uintptr_t in0) {
 
 /* NOTE: The following wrapper function is not covered by Coq proofs */
 static void fiat_secp256k1_nonzero(uint32_t* out1, const uint32_t arg1[8]) {
-  *out1 = internal_fiat_secp256k1_nonzero((uintptr_t)arg1);
+  *out1 = (uint32_t)internal_fiat_secp256k1_nonzero((uintptr_t)arg1);
 }
